test: move interceptors tests to its own file
