+++ jasync-sql/src/test/groovy/io/micronaut/configuration/jasync/ApplicationContextSpecification.groovy
@@ -1,0 +1,13 @@
+package io.micronaut.configuration.jasync
+
+import io.micronaut.context.ApplicationContext
+import spock.lang.AutoCleanup
+import spock.lang.Shared
+import spock.lang.Specification
+
+class ApplicationContextSpecification extends Specification implements ConfigurationFixture {
+
+    @AutoCleanup
+    @Shared
+    ApplicationContext applicationContext = ApplicationContext.run(configuration)
+}
