+++ jasync-sql/src/test/groovy/io/micronaut/configuration/jasync/ConfigurationFixture.groovy
@@ -1,0 +1,16 @@
+package io.micronaut.configuration.jasync
+
+trait ConfigurationFixture {
+
+    Map<String, Object> getConfiguration() {
+        Map<String, Object> m = [:]
+        if (specName) {
+            m['spec.name'] = specName
+        }
+        m
+    }
+
+    String getSpecName() {
+        null
+    }
+}
